move guideline hooks to a new file
